backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "passlib==1.7.4",
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-security-too>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "setuptools>=78.1.0",
    "argon2-cffi>=23.1.0",
    "flask-cors>=5.0.1",
    "redis>=5.2.1",
    "celery>=5.4.0",
    "flask-mail>=0.10.0",
    "python-dotenv>=1.1.0",
]

